add to cart item complete

+++ src/utility/index.js
@@ -4,7 +4,7 @@ const getAddedToCart = ()=> {
     const all = localStorage.getItem('addToCart')
     if(all){
         const addToCart = JSON.parse(all)
-        toast.success('Item Added Successfully!');
+        
         return addToCart
 
     }else{
@@ -15,8 +15,12 @@ const getAddedToCart = ()=> {
 const addCart = gadget => {
   
     const addToCart = getAddedToCart()
+    const isExist = addToCart.find(item => item.id == gadget.id)
+    if(isExist) return toast.error('This item already Exist!');
+
     addToCart.push(gadget)
     localStorage.setItem('addToCart', JSON.stringify(addToCart) )
+    toast.success('Item Added Successfully!');
 }
 
 export {addCart , getAddedToCart}
